SteveH: make JavaBean matchers implement Matcher<Object>
This is valid as the match method does in fact check the
actual class of the item being matched. It should help for cases where
the matcher of a concrete leaf in a type hierarchy is applied to an
item of the supertype -- e.g. using an AuthorIdeaRankingResultMatcher
to match an IdeaRankingResult that has not been downcast to
AuthorIdeaRankingResult